+++ common/src/main/java/dev/terminalmc/safevoid/SafeVoid.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2024 TerminalMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package dev.terminalmc.safevoid;
 
 import dev.terminalmc.safevoid.config.Config;
@@ -5,8 +21,6 @@ import dev.terminalmc.safevoid.util.ModLogger;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
 import net.minecraft.server.MinecraftServer;
-
-import static dev.terminalmc.safevoid.util.Localization.translationKey;
 
 public class SafeVoid {
     public static final String MOD_ID = "safevoid";
@@ -23,10 +37,8 @@ public class SafeVoid {
     }
 
     public static void onEndTick(MinecraftServer mc) {
-
     }
 
     public static void onConfigSaved(Config config) {
-        // If you are maintaining caches based on config values, update them here.
     }
 }