+++ common/src/main/java/dev/terminalmc/safevoid/command/Commands.java
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2024 TerminalMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package dev.terminalmc.safevoid.command;
 
 import com.mojang.brigadier.Command;
@@ -15,6 +31,7 @@ public class Commands<S> extends CommandDispatcher<S> {
     public void register(CommandDispatcher<S> dispatcher, CommandBuildContext buildContext,
                          net.minecraft.commands.Commands.CommandSelection selection) {
         dispatcher.register((LiteralArgumentBuilder<S>)literal(SafeVoid.MOD_ID)
+                .requires((sourceStack) -> sourceStack.hasPermission(2))
                 .then(literal("enable")
                         .executes(ctx -> {
                             MutableComponent msg = SafeVoid.PREFIX.copy();
@@ -25,7 +42,8 @@ public class Commands<S> extends CommandDispatcher<S> {
                                 Config.save();
                                 msg.append("Enabled");
                             }
-                            ctx.getSource().getPlayer().sendSystemMessage(msg);
+                            
+                            ctx.getSource().sendSystemMessage(msg);
                             return Command.SINGLE_SUCCESS;
                         })
                 )
@@ -39,7 +57,7 @@ public class Commands<S> extends CommandDispatcher<S> {
                                 Config.save();
                                 msg.append("Disabled");
                             }
-                            ctx.getSource().getPlayer().sendSystemMessage(msg);
+                            ctx.getSource().sendSystemMessage(msg);
                             return Command.SINGLE_SUCCESS;
                         })
                 )
@@ -48,7 +66,7 @@ public class Commands<S> extends CommandDispatcher<S> {
                             MutableComponent msg = SafeVoid.PREFIX.copy();
                             Config.loadAndSave();
                             msg.append("Config reloaded");
-                            ctx.getSource().getPlayer().sendSystemMessage(msg);
+                            ctx.getSource().sendSystemMessage(msg);
                             return Command.SINGLE_SUCCESS;
                         })
                 )